added value check

+++ src/main/java/app/services/implementation/AdminOperationsImpl.java
@@ -30,6 +30,11 @@ public class AdminOperationsImpl implements AdminOperations {
     public void addRoom() {
         ConsolePrinter.print("Enter new room name");
         String adminAnswerRoomName = reader.read();
+        Set<String> allRooms = placeDao.getAllRooms();
+        if (allRooms.contains(adminAnswerRoomName)) {
+            ConsolePrinter.print("This room already exists");
+            return;
+        }
         ConsolePrinter.print(placeDao.addNewPlace(adminAnswerRoomName, "room"));
     }
 
@@ -37,6 +42,11 @@ public class AdminOperationsImpl implements AdminOperations {
     public void addHall() {
         ConsolePrinter.print("Enter new hall name");
         String adminAnswerHallName = reader.read();
+        Set<String> allHalls = placeDao.getAllHalls();
+        if (!allHalls.contains(adminAnswerHallName)) {
+            ConsolePrinter.print("This hall already exists");
+            return;
+        }
         ConsolePrinter.print(placeDao.addNewPlace(adminAnswerHallName, "hall"));
     }
 
@@ -44,11 +54,15 @@ public class AdminOperationsImpl implements AdminOperations {
     public void addDesk() {
         ConsolePrinter.print("There are following rooms: ");
         Set<String> allRooms = placeDao.getAllRooms();
-        for (String room: allRooms){
+        for (String room : allRooms) {
             ConsolePrinter.print(room);
         }
         ConsolePrinter.print("Select the room where you want to add a desk");
         String adminAnswerRoom = reader.read();
+        if (!allRooms.contains(adminAnswerRoom)) {
+            ConsolePrinter.print("This room doesn't exist");
+            return;
+        }
         ConsolePrinter.print(placeDao.addNewDesk(adminAnswerRoom));
     }
 
@@ -56,42 +70,60 @@ public class AdminOperationsImpl implements AdminOperations {
     public void deleteRoom() {
         ConsolePrinter.print("There are following rooms: ");
         Set<String> allRooms = placeDao.getAllRooms();
-        for (String room: allRooms){
+        for (String room : allRooms) {
             ConsolePrinter.print(room);
         }
         ConsolePrinter.print("Select room name for deleting");
         String adminAnswerRoomName = reader.read();
+        if (!allRooms.contains(adminAnswerRoomName)) {
+            ConsolePrinter.print("This room doesn't exist");
+            return;
+        }
         ConsolePrinter.print(placeDao.deletePlace(adminAnswerRoomName));
     }
+
     @Override
     public void deleteDesk() {
         ConsolePrinter.print("There are following rooms: ");
         Set<String> allRooms = placeDao.getAllRooms();
-        for (String room: allRooms){
+        for (String room : allRooms) {
             ConsolePrinter.print(room);
         }
         ConsolePrinter.print("Select in which room you want to remove the desktop");
         String adminAnswerRoomName = reader.read();
+        if (!allRooms.contains(adminAnswerRoomName)) {
+            ConsolePrinter.print("This room doesn't exist");
+            return;
+        }
 
         ConsolePrinter.print("There are following desks: ");
         Set<Integer> allDesksInRoom = placeDao.getSetOfAllDesksInRoom(adminAnswerRoomName);
-        for (int desk: allDesksInRoom){
+        for (int desk : allDesksInRoom) {
             ConsolePrinter.print("Desk № " + desk);
         }
         ConsolePrinter.print("Select the table number you want to delete");
         String adminAnswerDeskNumber = reader.read();
+        if (!allDesksInRoom.contains(Integer.parseInt(adminAnswerDeskNumber))) {
+            ConsolePrinter.print("This desk doesn't exist");
+            return;
+        }
 
         ConsolePrinter.print(placeDao.deleteDesk(Integer.parseInt(adminAnswerDeskNumber), adminAnswerRoomName));
     }
+
     @Override
     public void deleteHall() {
         ConsolePrinter.print("There are following hall: ");
         Set<String> allHalls = placeDao.getAllHalls();
-        for (String hall: allHalls){
+        for (String hall : allHalls) {
             ConsolePrinter.print(hall);
         }
         ConsolePrinter.print("Select hall name for deleting");
         String adminAnswerHallName = reader.read();
+        if (!allHalls.contains(adminAnswerHallName)) {
+            ConsolePrinter.print("This hall doesn't exist");
+            return;
+        }
         ConsolePrinter.print(placeDao.deletePlace(adminAnswerHallName));
     }
